Move new classes to commons
