+++ universal-login-commons/src/core/constants/isStringANumberRegex.ts
