+++ universal-login-commons/src/core/utils/ValuePresenter.ts
@@ -1,10 +1,10 @@
-import {isStringANumber} from '@unilogin/commons';
+import {isNumber} from './isNumber';
 
 export class ValuePresenter {
   static presentRoundedValue(value: string, roundFunc: (value: string, precision?: number) => string, precision?: number): string {
     if (value === '') {
       return '';
-    } else if (!isStringANumber(value)) {
+    } else if (!isNumber(value)) {
       return value;
     }
     return roundFunc(value, precision === undefined ? 3 : precision);
