+++ universal-login-commons/src/core/utils/ValueRounder.ts
@@ -1,8 +1,8 @@
-import {isStringANumber} from '@unilogin/commons';
+import {isNumber} from './isNumber';
 
 export class ValueRounder {
   static ceil(value: string, precision = 3): string {
-    if (!isStringANumber(value)) {
+    if (!isNumber(value)) {
       throw Error('String is not a number');
     }
     const roundLimit = Math.pow(10, precision);
@@ -10,7 +10,7 @@ export class ValueRounder {
   }
 
   static floor(value: string, precision = 3): string {
-    if (!isStringANumber(value)) {
+    if (!isNumber(value)) {
       throw Error('String is not a number');
     }
     const roundLimit = Math.pow(10, precision);
