+++ universal-login-commons/src/core/utils/isNumber.ts
@@ -1,5 +1,5 @@
 import {Result, asNumber} from '@restless/sanitizers';
 
-export const isStringANumber = (value: string): boolean => {
+export const isNumber = (value: string): boolean => {
   return Result.isOk(asNumber(value, ''));
 };
