+++ universal-login-commons/test/core/utils/ValueRounder.test.ts
@@ -1,5 +1,5 @@
 import {expect} from 'chai';
-import {ValueRounder} from '../../src/app/valueRounder';
+import {ValueRounder} from '../../../src';
 
 describe('UNIT: valueRounder', () => {
   describe('Ceil', () => {
