+++ universal-login-react/src/ui/commons/Assets.tsx
@@ -1,14 +1,13 @@
 import React, {useState} from 'react';
 import {utils} from 'ethers';
 import {DeployedWallet} from '@unilogin/sdk';
-import {TokenDetailsWithBalance} from '@unilogin/commons';
+import {TokenDetailsWithBalance, ValueRounder} from '@unilogin/commons';
 import {Asset} from './Asset';
 import './../styles/assetsList.sass';
 import './../styles/assetsListDefaults.sass';
 import {useAsyncEffect} from '../hooks/useAsyncEffect';
 import {getStyleForTopLevelComponent} from '../../core/utils/getStyleForTopLevelComponent';
 import {getIconForToken} from '../../core/utils/getIconForToken';
-import {ValueRounder} from '../../app/valueRounder';
 
 export interface AssetsProps {
   deployedWallet: DeployedWallet;
